+++ src/main/java/com/supergt/racingdriver/RacingdriverApplication.java
@@ -5,8 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class RacingdriverApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RacingdriverApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(RacingdriverApplication.class, args);
+    }
 }
