--- src/main/java/com/supergt/racingdriver/response/RacerCreateResponse.java
@@ -1,13 +1,0 @@
-package com.supergt.racingdriver.response;
-
-public class RacerCreateResponse {
-    private String message;
-
-    public RacerCreateResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
