--- src/main/java/com/supergt/racingdriver/response/RacerDeleteResponse.java
@@ -1,14 +1,0 @@
-package com.supergt.racingdriver.response;
-
-public class RacerDeleteResponse {
-
-    public String message;
-
-    public RacerDeleteResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
