--- src/main/java/com/supergt/racingdriver/response/RacerQueryResponse.java
@@ -1,13 +1,0 @@
-package com.supergt.racingdriver.response;
-
-public class RacerQueryResponse {
-    public String message;
-
-    public RacerQueryResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
