+++ src/main/java/com/supergt/racingdriver/response/RacerResponse.java
@@ -3,13 +3,9 @@ package com.supergt.racingdriver.response;
 import java.time.LocalDate;
 
 public class RacerResponse {
-
     private String givenName;
-
     public String familyName;
-
     public LocalDate birthday;
-
 
     public RacerResponse(String givenName, String familyName, LocalDate birthday) {
         this.givenName = givenName;
@@ -27,5 +23,35 @@ public class RacerResponse {
 
     public LocalDate getBirthday() {
         return birthday;
+    }
+
+    public class Response {
+        private String message;
+
+        public Response(String message) {
+            this.message = message;
+        }
+
+        public String getMessage() {
+            return message;
+        }
+    }
+
+    public class RacerCreateResponse extends Response {
+        public RacerCreateResponse(String message) {
+            super(message);
+        }
+    }
+
+    public class RacerUpdateResponse extends Response {
+        public RacerUpdateResponse(String message) {
+            super(message);
+        }
+    }
+
+    public class RacerDeleteResponse extends Response {
+        public RacerDeleteResponse(String message) {
+            super(message);
+        }
     }
 }
