--- src/main/java/com/supergt/racingdriver/response/RacerUpdateResponse.java
@@ -1,14 +1,0 @@
-package com.supergt.racingdriver.response;
-
-public class RacerUpdateResponse {
-
-    public String message;
-
-    public RacerUpdateResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
